Refactor project structure to use Common module for shared resources
- Updated import statements across multiple files to reference the new Common module for Tools, InfomationType, and StockInfoData.
- Created new Common module files: Globals.py, InfomationType.py, Parameter.py, StockInfoData.py, and Tools.py to centralize common functionalities and data structures.
