+++ BackTestService/__init__.py
@@ -1,1 +1,1 @@
-from .BackTestStock import BackTestStock
+from .BacktestFilter import *
