+++ Common/Globals.py
