+++ Common/StockInfoData.py
@@ -10,7 +10,7 @@ class StockInfoData:
         self.name: str = name
         self.number: int = int(number)
         self.type: str = type
-        self.start: datetime = datetime.strptime(start, "%Y/%m/%d")
+        self.start: datetime = start
         self.market: str = market
         self.group: str = group
