+++ MongoService.py
@@ -5,7 +5,7 @@ import pymongo
 import twstock  # 抓取台灣股票資料套件
 from pymongo import MongoClient, database
 
-import Tools
+import Common.Tools as Tools
 
 
 class MongoService:
