+++ UnitTest/BackTestInfoData_test.py
@@ -3,7 +3,7 @@ from datetime import datetime
 
 from BackTestService.BackTestInfoData import BackTestInfoDataPriceByToday
 from FilterService.StockHistory import OriginalStockTest
-from StockInfoData import BaseInfoData
+from Common.StockInfoData import BaseInfoData
 
 
 class BackTestInfoDataPriceByToday_test(unittest.TestCase):
